Thesis modelling, and metadata mixed bag:
* share properties with `Collection`/`Community`/`Item`/`Thesis` via concern;
* handling shared properties with `Item` via concern;
* create attributes distinct to `Thesis`.
* rename `Item` property `derived_from` to `source`;
* solr index for `item_type_with_status` shared with `Item` for facetting;
* derived property `sort_year` derived from graduation date for `Thesis`;
* `bibo:status` provided by `rdf-vocab`, so remove from `terms.yml`
* move `Community` to `ual` vocabulary namespace (uppercase intentional);
* remove fake `jupiter_core` vocabulary namespace.
