+++ app/models/item.rb
@@ -1,28 +1,17 @@
-
 class Item < JupiterCore::LockedLdpObject
 
-  ldp_object_includes Hydra::Works::WorkBehavior
-
-  VISIBILITY_EMBARGO = CONTROLLED_VOCABULARIES[:visibility].embargo.freeze
-  VISIBILITIES = (JupiterCore::VISIBILITIES + [VISIBILITY_EMBARGO]).freeze
-  VISIBILITIES_AFTER_EMBARGO = [CONTROLLED_VOCABULARIES[:visibility].authenticated,
-                                CONTROLLED_VOCABULARIES[:visibility].draft,
-                                CONTROLLED_VOCABULARIES[:visibility].public].freeze
+  include CommonObjectProperties
+  include CommonItemProperties
 
   ALLOWED_LICENSES = (CONTROLLED_VOCABULARIES[:license] + CONTROLLED_VOCABULARIES[:old_license]).freeze
-
-  # Dublin Core attributes
-  has_attribute :title, ::RDF::Vocab::DC.title, solrize_for: [:search, :sort]
 
   # Contributors (faceted in `all_contributors`)
   has_multival_attribute :creators, ::RDF::Vocab::DC11.creator, solrize_for: [:search]
   has_multival_attribute :contributors, ::RDF::Vocab::DC11.contributor, solrize_for: [:search]
 
   has_attribute :created, ::RDF::Vocab::DC.created, solrize_for: [:search, :sort]
-  has_attribute :sort_year, ::TERMS[:ual].sortyear, solrize_for: [:search, :sort, :facet]
 
   # Subject types (see `all_subjects` for faceting)
-  has_multival_attribute :subject, ::RDF::Vocab::DC11.subject, solrize_for: [:search]
   has_multival_attribute :temporal_subjects, ::RDF::Vocab::DC.temporal, solrize_for: [:search]
   has_multival_attribute :spatial_subjects, ::RDF::Vocab::DC.spatial, solrize_for: [:search]
 
@@ -30,36 +19,17 @@ class Item < JupiterCore::LockedLdpObject
   has_attribute :publisher, ::RDF::Vocab::DC.publisher, solrize_for: [:search, :facet]
   # has_attribute :date_modified, ::RDF::Vocab::DC.modified, type: :date, solrize_for: :sort
   has_multival_attribute :languages, ::RDF::Vocab::DC.language, solrize_for: [:search, :facet]
-  has_attribute :embargo_end_date, ::RDF::Vocab::DC.available, type: :date, solrize_for: [:sort]
   has_attribute :license, ::RDF::Vocab::DC.license, solrize_for: [:search]
-  has_attribute :rights, ::RDF::Vocab::DC11.rights, solrize_for: :exact_match
+
   # `type` is an ActiveFedora keyword, so we call it `item_type`
   # Note also the `item_type_with_status` below for searching, faceting and forms
   has_attribute :item_type, ::RDF::Vocab::DC.type, solrize_for: :exact_match
-  has_attribute :derived_from, ::RDF::Vocab::DC.source, solrize_for: :exact_match
-  has_multival_attribute :is_version_of, ::RDF::Vocab::DC.isVersionOf, solrize_for: :exact_match
-  has_attribute :alternative_title, ::RDF::Vocab::DC.alternative, solrize_for: :search
+  has_attribute :source, ::RDF::Vocab::DC.source, solrize_for: :exact_match
   has_attribute :related_link, ::RDF::Vocab::DC.relation, solrize_for: :exact_match
-
-  # UAL attributes
-  has_attribute :depositor, ::TERMS[:ual].depositor, solrize_for: [:search]
-  has_attribute :fedora3_handle, ::TERMS[:ual].fedora3handle, solrize_for: :exact_match
-  has_attribute :fedora3_uuid, ::TERMS[:ual].fedora3uuid, solrize_for: :exact_match
-  has_attribute :ingest_batch, ::TERMS[:ual].ingestbatch, solrize_for: :exact_match
-  has_multival_attribute :member_of_paths, ::TERMS[:ual].path,
-                         type: :path,
-                         solrize_for: :pathing
-
-  # Prism attributes
-  has_attribute :doi, ::TERMS[:prism].doi, solrize_for: :exact_match
 
   # Bibo attributes
   # This status is only for articles: either 'published' (alone) or two triples for 'draft'/'submitted'
-  has_multival_attribute :publication_status, ::TERMS[:bibo].status, solrize_for: :exact_match
-
-  # Project Hydra ACL attributes
-  has_multival_attribute :embargo_history, ::TERMS[:acl].embargoHistory, solrize_for: :exact_match
-  has_attribute :visibility_after_embargo, ::TERMS[:acl].visibilityAfterEmbargo, solrize_for: :exact_match
+  has_multival_attribute :publication_status, ::RDF::Vocab::BIBO.status, solrize_for: :exact_match
 
   # Solr only
   additional_search_index :doi_without_label, solrize_for: :exact_match,
@@ -71,20 +41,12 @@ class Item < JupiterCore::LockedLdpObject
                           solrize_for: :facet,
                           as: -> { item_type_with_status_code }
 
-  # Combine creators and contributors for faceting
+  # Combine creators and contributors for faceting (Thesis also uses this index)
   # Note that contributors is converted to an array because it can be nil
   additional_search_index :all_contributors, solrize_for: :facet, as: -> { creators + contributors.to_a }
 
   # Combine all the subjects for faceting
   additional_search_index :all_subjects, solrize_for: :facet, as: -> { all_subjects }
-
-  def self.display_attribute_names
-    super - [:member_of_paths]
-  end
-
-  def self.valid_visibilities
-    super + [VISIBILITY_EMBARGO]
-  end
 
   # This is stored in solr: combination of item_type and publication_status
   def item_type_with_status_code
@@ -101,126 +63,27 @@ class Item < JupiterCore::LockedLdpObject
     return nil
   end
 
-  def doi_url
-    "https://doi.org/#{read_solr_index(:doi_without_label).first}"
-  end
-
   def all_subjects
     subject + temporal_subjects.to_a + spatial_subjects.to_a
   end
 
-  def file_sets
-    FileSet.where(item: id)
-  end
-
-  def each_community_collection
-    member_of_paths.each do |path|
-      community_id, collection_id = path.split('/')
-      yield Community.find(community_id), Collection.find(collection_id)
-    end
-  end
-
-  # TODO: implement me
-  def thumbnail
-    nil
-  end
-
   unlocked do
-    validates :embargo_end_date, presence: true, if: ->(item) { item.visibility == VISIBILITY_EMBARGO }
-    validates :embargo_end_date, absence: true, if: ->(item) { item.visibility != VISIBILITY_EMBARGO }
-    validates :visibility_after_embargo, presence: true, if: ->(item) { item.visibility == VISIBILITY_EMBARGO }
-    validates :visibility_after_embargo, absence: true, if: ->(item) { item.visibility != VISIBILITY_EMBARGO }
-    validates :member_of_paths, presence: true
-    validates :title, presence: true
     validates :languages, presence: true
     validates :item_type, presence: true
     validates :subject, presence: true
     validates :creators, presence: true
-    validate :communities_and_collections_validations
     validate :language_validations
     validate :license_and_rights_validations
-    validate :visibility_after_embargo_validations
     validate :item_type_and_publication_status_validations
 
     before_validation do
-      # TODO: for theses, the sort_year attribute should be derived from ual:graduationDate
+      # Note: for Thesis, the sort_year attribute is derived from ual:graduationDate
       begin
         self.sort_year = Date.parse(created).year.to_s if created.present?
       rescue ArgumentError
         # date was unparsable, try to pull out the first 4 digit number as a year
         capture = created.scan(/\d{4}/)
         self.sort_year = capture[0] if capture.present?
-      end
-    end
-
-    before_save do
-      if member_of_paths_changed?
-        # This adds the `pcdm::memberOf` predicates, pointing to each collection
-        self.member_of_collections = []
-        member_of_paths.each do |path|
-          _community_id, collection_id = path.split('/')
-          collection = Collection.find_by(collection_id)
-
-          # TODO: can this be streamlined so that a fetch from Fedora isn't needed?
-          collection.unlock_and_fetch_ldp_object do |unlocked_collection|
-            self.member_of_collections += [unlocked_collection]
-          end
-        end
-      end
-    end
-
-    def communities_and_collections_validations
-      return if member_of_paths.blank?
-      member_of_paths.each do |path|
-        community_id, collection_id = path.split('/')
-        community = Community.find_by(community_id)
-        errors.add(:member_of_paths, :community_not_found, id: community_id) if community.blank?
-        collection = Collection.find_by(collection_id)
-        errors.add(:member_of_paths, :collection_not_found, id: collection_id) if collection.blank?
-      end
-    end
-
-    def add_to_path(community_id, collection_id)
-      self.member_of_paths += ["#{community_id}/#{collection_id}"]
-      # TODO: also add the collection (not the community) to the Item's memberOf relation, as metadata
-      # wants to continue to model this relationship in pure PCDM terms, and member_of_path is really for our needs
-      # so that we can facet by community and/or collection properly
-      # TODO: add collection_id to member_of_collections
-    end
-
-    def add_communities_and_collections(communities, collections)
-      return unless communities.present? && collections.present?
-      communities.each_with_index do |community, idx|
-        add_to_path(community, collections[idx])
-      end
-    end
-
-    def add_files(files)
-      return if files.blank?
-      # Need a item id for file sets to point to
-      # TODO should this be a side effect? should we throw an exception if there's no id? Food for thought
-      save! if id.nil?
-
-      files.each do |file|
-        FileSet.new_locked_ldp_object.unlock_and_fetch_ldp_object do |unlocked_fileset|
-          unlocked_fileset.owner = owner
-          unlocked_fileset.visibility = visibility
-          Hydra::Works::AddFileToFileSet.call(unlocked_fileset, file, :original_file,
-                                              update_existing: false, versioning: false)
-          unlocked_fileset.member_of_collections += [self]
-          # Temporarily cache the file name for storing in Solr
-          # if the file was uploaded, it responds to +original_filename+
-          # if it's a Ruby File object, it has a +basename+. This distinction seems arbitrary.
-          unlocked_fileset.contained_filename = if file.respond_to?(:original_filename)
-                                                  file.original_filename
-                                                else
-                                                  File.basename(file)
-                                                end
-          unlocked_fileset.save!
-          self.members += [unlocked_fileset]
-          # pull in hydra derivatives, set temp file base
-          # Hydra::Works::CharacterizationService.run(fileset.characterization_proxy, filename)
-        end
       end
     end
 
@@ -241,12 +104,6 @@ class Item < JupiterCore::LockedLdpObject
         end
         errors.add(:base, :not_both_license_and_rights) if rights.present?
       end
-    end
-
-    def visibility_after_embargo_validations
-      return if visibility_after_embargo.nil?
-      return if VISIBILITIES_AFTER_EMBARGO.include?(visibility_after_embargo)
-      errors.add(:visibility_after_embargo, :not_recognized)
     end
 
     def item_type_and_publication_status_validations
