+++ config/controlled_vocabularies/institution.yml
@@ -1,0 +1,6 @@
+---
+insitutution:
+  - code: uofa
+    uri: http://id.loc.gov/authorities/names/n79058482
+  - code: st_stephens
+    uri: http://id.loc.gov/authorities/names/n2009054054
