+++ config/terms.yml
@@ -18,25 +18,29 @@
 - vocabulary: ual
   schema: http://terms.library.ualberta.ca/
   terms:
+    - committeeMember
+    - Community # rdf:type for community objects, Zach has advised uppercase for class
     - depositor
+    - dissertant
     - fedora3handle
     - fedora3uuid
+    - graduationDate
+    - hydraNoid
     - ingestbatch
     - path
-    - sortyear
-    - hydraNoid
+    - proquest
     # separate from date_created, we need to track a precise time the metadata record was created
     - recordCreatedInJupiter
+    - sortyear
+    - specialization
+    - supervisor
+    - thesisLevel
+    - unicorn
 
 - vocabulary: prism
   schema: http://prismstandard.org/namespaces/basic/3.0/
   terms:
     - doi
-
-- vocabulary: bibo
-  schema: http://purl.org/ontology/bibo/
-  terms:
-    - status
 
 - vocabulary: acl
   schema: 'http://projecthydra.org/ns/auth/acl#'
@@ -44,7 +48,7 @@
     - embargoHistory
     - visibilityAfterEmbargo
 
-- vocabulary: jupiter_core
-  schema: 'http://fake.figurethisout.todo#'
+- vocabulary: swrc
+  schema: 'http://ontoware.org/swrc/ontology#'
   terms:
-    - community # rdf:type for community objects, which aren't merely rdf:objects but our own custom derivation
+    - institution
